Update to Idris rover glb model
I've added the corrected texture for the Aber sign (Thank you to Sivert for the photo) and added another sign on the other side of the rover.
I've also added the four aerials that were missing from the upper frame. I know there's a metal plate missing on the sensor mast, but no one has noticed yet :smirk: