Adding of the iCub glb and blender file
I've added the missing pieces for the iCub and put it in a nice pose for the glb file. The Blender file is included, but please note that the environment texture is externally referenced. So if you want to make some renders you'll have to find a replacement HDR map.
I got mine at HDRI Haven https://hdrihaven.com/hdris/